# optimize simple_solution_generator()