this is a somewhat expensive part of adding a transaction to the mempool. With the planned raise of the mempool size this is an attempt to mitigate load- and performance issues caused by higher pressure on the mempool.

The benchmark shows a small improvement on wall-clock time, but it's CPU time in the main process that's the more precious resource. We already limit validation to 2 threads.

CPU usage:

| CPU usage | before | after | after / before |
| --- | --- | --- | --- |
| mempool_manager.add_spendbundle() | 13.58% | 6.12% | 0.45 |

main:
```
add_spendbundle time: 2.1495s
2.15ms per add_spendbundle() call
```

this patch:
```
add_spendbundle time: 2.0407s
2.04ms per add_spendbundle() call
```

# profiles

main:

![mempool-add-mt](https://user-images.githubusercontent.com/661450/183031243-5a01751a-13ab-4ffb-afd5-1915b770de16.png)

this patch:

![mempool-add-mt](https://user-images.githubusercontent.com/661450/183031096-cfd63a34-2c6e-4ed9-b83b-daa719956d5c.png)