@mariano54 2022-08-05
Nice!

We also call `simple_solution_generator(new_spend).program` in the prevalidation, so maybe we should send that over to not have to recompute it.