@arvidn 2022-08-05
> We also call simple_solution_generator(new_spend).program in the prevalidation, so maybe we should send that over to not have to recompute it.

Yeah, I was considering that patch also. but after making this change it disappeared from the profile, so I'm thinking it might be a wash of sending it over the pipe or regenerating it